Added the basic framework for files
These added files are the basis for what the project will contain. I
intend for there to be a character class that contains all the basic
operations for the character sheet. The character functions are intended
to allow for editing of the character class and doing operations with
it. The ch_vector files are intended to be used to read a list of saved
files into the program and then allow the user to access which file they
would like to use.