Added makefile and other files.
Added and implemented the makefile. created some files to store the
inventory of the character to be implemented with an item class that
will contain the item type(weapon, object, food, valuable, etc.) the
weight, the price, the stats, the name, and the description of each item. This will
allow for the creation of custom items and the addition of any other
currenty existing item.