Added implementations for some setters
item.h now has the containers for the setter functions as well as the
implementation for the constructor for the program. This will simply set
the values of the item object based on the parameters passed. Still need
to figure out how to store the item vectors in a file and read them into
an item object easily.