Wrote the header file for the weapon class
Not sure why we are doing inheritance for this now. The item parameters
are different enough that I think having a weapon inherit from an item
might be a little goofy. There are elements exclusive to the weapon that
just make inherited classes more complicated than the class above it. At
least how they are currently implemented.